requirements.txt
langchain
streamlit
faiss-cpu
python-dotenv
tiktoken
together
fireworks
sentence_transformers
langchain-community
PyPDF2
python-docx